# Conflicts:
#	.github/workflows/ant.yml
#	classes/classes/Scenes/Areas/Beach/ArigeanOmnibusAbomination.as
#	classes/classes/Scenes/Combat/AbstractBloodSoulSkill.as
#	classes/classes/Scenes/Combat/AbstractSoulSkill.as
#	classes/classes/Scenes/Combat/General/FlyingSwordSkill.as
#	classes/classes/Scenes/Combat/General/GoblinLustBombSkill.as
#	classes/classes/Scenes/Combat/General/MummyAttackSkill.as
#	classes/classes/Scenes/Combat/General/TeaseSkill.as
#	classes/classes/Scenes/Combat/General/WillOfTheWispSkill.as
#	classes/classes/Scenes/Combat/Soulskills/BladeDanceSkill.as
#	classes/classes/Scenes/Combat/Soulskills/BladeHailSkill.as
#	classes/classes/Scenes/Combat/Soulskills/BloodDewdropsPuppiesSkill.as
#	classes/classes/Scenes/Combat/Soulskills/BloodDewdropsSkill.as
#	classes/classes/Scenes/Combat/Soulskills/BloodReqiuemSkill.as
#	classes/classes/Scenes/Combat/Soulskills/BloodSwipePuppiesSkill.as
#	classes/classes/Scenes/Combat/Soulskills/BloodSwipeSkill.as
#	classes/classes/Scenes/Combat/Soulskills/CleansingPalmSkill.as
#	classes/classes/Scenes/Combat/Soulskills/CometSkill.as
#	classes/classes/Scenes/Combat/Soulskills/CreateElementSkill.as
#	classes/classes/Scenes/Combat/Soulskills/DevourerSkill.as
#	classes/classes/Scenes/Combat/Soulskills/DracoSweepSkill.as
#	classes/classes/Scenes/Combat/Soulskills/EarthStanceSkill.as
#	classes/classes/Scenes/Combat/Soulskills/FingerOfDeathSkill.as
#	classes/classes/Scenes/Combat/Soulskills/FirePunchSkill.as
#	classes/classes/Scenes/Combat/Soulskills/FlamesOfLoveSkill.as
#	classes/classes/Scenes/Combat/Soulskills/HeartSeekerPuppiesSkill.as
#	classes/classes/Scenes/Combat/Soulskills/HeartSeekerSkill.as
#	classes/classes/Scenes/Combat/Soulskills/HurricaneDanceSkill.as
#	classes/classes/Scenes/Combat/Soulskills/IceFistSkill.as
#	classes/classes/Scenes/Combat/Soulskills/IciclesOfLoveSkill.as
#	classes/classes/Scenes/Combat/Soulskills/KillingIntentSkill.as
#	classes/classes/Scenes/Combat/Soulskills/ManyBirdsSkill.as
#	classes/classes/Scenes/Combat/Soulskills/MultiThrustSkill.as
#	classes/classes/Scenes/Combat/Soulskills/NightOfBrotherhoodSkill.as
#	classes/classes/Scenes/Combat/Soulskills/PunishingKickSkill.as
#	classes/classes/Scenes/Combat/Soulskills/ScarletSpiritChargeSkill.as
#	classes/classes/Scenes/Combat/Soulskills/SoulBlastSkill.as
#	classes/classes/Scenes/Combat/Soulskills/SoulDrainSkill.as
#	classes/classes/Scenes/Combat/Soulskills/StormOfSisterhoodSkill.as
#	classes/classes/Scenes/Combat/Soulskills/TranceSkill.as
#	classes/classes/Scenes/Combat/SpecialsMagic/EAspectWaterSkill.as
#	classes/classes/Scenes/Combat/SpecialsMagic/EAspectWoodSkill.as
#	classes/classes/Scenes/Combat/SpecialsMagic/FusionSpecial2ndSkill.as
#	classes/classes/Scenes/Dungeons/EbonLabyrinth/Draculina.as
#	classes/classes/Scenes/Dungeons/EbonLabyrinth/DraculinaScene.as

+++ classes/classes/Scenes/Combat/Soulskills/KillingIntentSkill.as
@@ -15,13 +15,13 @@ public class KillingIntentSkill extends AbstractSoulSkill {
             "Project your bloodlust through your soul sense, damaging the enemies' mind. Damage increases with the number of monsters killed and cultivation rank.",
             TARGET_ENEMY,
             TIMING_INSTANT,
-            [TAG_DAMAGING, TAG_MAGICAL],
+            [TAG_DAMAGING, TAG_MAGICAL, TAG_TIER1],
             PerkLib.KillingIntent
         )
-		baseSFCost = 50;
+		baseSFCost = 200;
 		lastAttackType = Combat.LAST_ATTACK_SPELL;
     }
-
+/*
     override protected function usabilityCheck():String {
         var uc:String =  super.usabilityCheck();
         if (uc) return uc;
@@ -32,46 +32,25 @@ public class KillingIntentSkill extends AbstractSoulSkill {
 
         return "";
     }
-
+*/
 	override public function describeEffectVs(target:Monster):String {
 		return "~" + numberFormat(calcDamage(target)) + " true damage. Can cause fear";
 	}
 
-	private function calcDeathCap():Number {
-		var cap:Number = 0;
-
-		if (player.hasPerk(PerkLib.SoulApprentice)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulPersonage)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulWarrior)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulSprite)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulScholar)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulGrandmaster)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulElder)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulExalt)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulOverlord)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulTyrant)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulKing)) cap += 10;
-		if (player.hasPerk(PerkLib.SoulEmperor)) cap += 10;
-    
-		return cap;
-	}
-
 	private function calcDeathMod():Number {
 		var mod:Number = 1;
-		var deathCap:Number = calcDeathCap();
-		var monstersKilled:int = 	flags[kFLAGS.GOBLINS_KILLED] + 
+		var monstersKilled:int = 	flags[kFLAGS.THIEFS_KILLED] +
+									flags[kFLAGS.GOBLINS_KILLED] + 
 									flags[kFLAGS.HELLHOUNDS_KILLED] +
 									flags[kFLAGS.IMPS_KILLED] +
 									flags[kFLAGS.MINOTAURS_KILLED] +
 									flags[kFLAGS.TRUE_DEMONS_KILLED];
-		if (deathCap != -1) monstersKilled = Math.min(deathCap, monstersKilled);
-		mod += 0.1 * monstersKilled;
+		mod += 0.1 * flags[kFLAGS.SOUL_CULTIVATION] * monstersKilled;
 		return mod;
 	}
 
 	override public function calcCooldown():int {
-		var cooldown:int = 3;
-		return cooldown;
+		return soulskillCooldown(3, false);
 	}
 
 	public function calcDamage(monster:Monster):Number {
@@ -83,28 +62,24 @@ public class KillingIntentSkill extends AbstractSoulSkill {
 		//other bonuses
 		if (player.hasPerk(PerkLib.Heroism) && (monster && (monster.hasPerk(PerkLib.EnemyBossType) || monster.hasPerk(PerkLib.EnemyHugeType)))) damage *= 2;
 		if (player.perkv1(IMutationsLib.AnubiHeartIM) >= 4 && player.HP < Math.round(player.maxHP() * 0.5)) damage *= 1.5;
-
 		return Math.round(damage);
 	}
 
     override public function doEffect(display:Boolean = true):void {
 		clearOutput();
 		if (monster.hasPerk(PerkLib.Insanity) || monster.inte == 0) {
-			if (display) outputText("You try to strike the enemies' resolve with your bloodlust, but the monater is unaffected. It doesn't seem as though bloodlust registers to them!\n\n");
+			if (display) outputText("You try to strike the enemy's resolve with your bloodlust, but the monster is unaffected. It doesn't seem as though bloodlust registers to them!\n\n");
 			return;
 		} else if (monster is LivingStatue) {
 			if (display) outputText("Bloodlust does not matter to a statue!\n\n");
 			return;
 		}
-
 		if (silly()) 
 			if (display) outputText("You narrow your eyes at this foolish upstart. \"<i>Junior, you dare!</i>\" Your power of your soul flares from your body, seeking to strike them down. ");
 		else {
-			outputText("You set your focus on [themonster], sending your desire to kill them through your soulforce. ");
+			if (display) outputText("You set your focus on [themonster], sending your desire to kill them through your soulforce. ");
 		}
-
 		var damage:Number = calcDamage(monster);
-
 		//Determine if critical hit!
 		var crit:Boolean = false;
 		var critChance:int = 5;
@@ -114,18 +89,13 @@ public class KillingIntentSkill extends AbstractSoulSkill {
 			crit = true;
 			damage *= 1.75;
 		}
-		
-		
 		if (display) outputText("Your intent slams into [themonster], sending them reeling from your bloodlust! ");
 		doTrueDamage(damage, true, display);
-		
 		if (crit && display) outputText(" <b>*Critical Hit!*</b>");
-
 		if (rand(5) == 0 && monster.level <= player.level) {
 			monster.createStatusEffect(StatusEffects.Fear,1,0,0,0);
 			if (display) outputText(" [Themonster] is left cowed from your mental attack!");
 		}
-
 		checkAchievementDamage(damage);
 		if (display) outputText("\n\n");
 		if (!player.hasStatusEffect(StatusEffects.BloodCultivator) && flags[kFLAGS.IN_COMBAT_PLAYER_ANUBI_HEART_LEECH] == 0) anubiHeartLeeching(damage);
